MVP2.1.0.2: 3 backlog bug fixes + 2.1.1 design notes
Three user-discovered bugs (all pre-2.1.0), all fixed:

Bug fixes (all from doc/v2.1.0.2-release-notes.md):
- Video.duration column was Integer but stores floats (336.44 →
  336 silently). Changed to Float. No migration needed (SQLite
  is dynamically typed; existing rows stay valid).
- Video.file_size was NOT updated on swap (after a 54 MB WebM
  → 11 MB MP4 swap, DB still showed 54 MB). Now stat()s the new
  file and stamps it before commit. Also adds both sizes to the
  audit log message ('Swapped from lesson.webm (56.6 MB) to
  v1.mp4 (11.0 MB). Transcript and materials preserved.').
- get_video_file hardcoded Content-Type: video/mp4 regardless
  of file extension. Now maps .mp4 / .webm / .avi / .mov /
  .mkv / .m4v to their correct MIME types, with
  application/octet-stream fallback for unknown extensions.

Bonus fix (not a bug, but spotted while testing):
- extra_json was stored as Python str({...}) (single-quoted
  repr) instead of json.dumps(...) (double-quoted JSON). Fixed
  in both places: run_plugin and swap_video_file_to. External
  consumers can now parse it as JSON.

Stale docs: MVP2.1-Status.md, MVP2.1-all.md, Readme.md
updated to reflect 2.1.0.1 + 2.1.0.2 + 2.1.1 (deferred).

2.1.1 design notes appended to MVP2.1-Status.md §3 for
future reference (per user request: 'write this down
somewhere for future reference'). The plugin pool shipped
in 2.1.0.1 is the foundation; the general pool for upload /
transcribe / generate uses the same asyncio.Queue +
asyncio.Semaphore(N) pattern, with one separate pool per
category (per user preference documented in 2.1.0.1 §6 —
separate pools with their own limits, not one shared pool).
Pool limits: upload 1-2, transcribe 1, generate 1, plugin 3.

Tests: 623 passing (was 614, +9 new tests; 0 regressions;
1 pre-existing whisper-picker flake unrelated).
9 new tests in tests/test_backlog_bugs_2_1_0_2.py:
  1. test_video_duration_column_is_float
  2. test_video_duration_stores_float_value
  3. test_swap_updates_file_size
  4. test_swap_audit_log_includes_size_info
  5-8. test_get_video_file_returns_correct_mime_for_{mp4,webm,mov,mkv}
  9. test_get_video_file_unknown_extension_falls_back

+++ app/models/video.py
@@ -74,7 +74,15 @@ class Video(Base):
     filename: Mapped[str] = mapped_column(String(512), nullable=False)
     file_path: Mapped[str] = mapped_column(String(1024), nullable=False)
     file_size: Mapped[int] = mapped_column(Integer, default=0)
-    duration: Mapped[float] = mapped_column(Integer, default=0)  # seconds
+    # MVP2.1.0.2 — `duration` was declared as Integer but stores floats
+    # like 336.44 (Whisper's segment-end timestamps are floats with
+    # sub-second precision; rounding to int loses the millisecond
+    # accuracy in the course-page badge). Changed to Float. The DB
+    # column is now REAL (SQLite) / DOUBLE PRECISION (Postgres), and
+    # `Base.metadata.create_all()` will use the new type for new
+    # tables. Existing rows keep their integer values (SQLite will
+    # happily cast int → float on read).
+    duration: Mapped[float] = mapped_column(Float, default=0.0)  # seconds
     order_index: Mapped[int] = mapped_column(Integer, default=0)
     section_id: Mapped[str] = mapped_column(
         String(36), ForeignKey("sections.id", ondelete="CASCADE"), nullable=False
